# Optimizing grpc_error Part 3: const char* -> grpc_slice