Based off error opt part1 (#9756), part2 (#9778), and also #9745.

This is the start of the PR that plumbs grpc_slice through all of the code that uses grpc_errors. For now I only changes the two inlined strings, file and desc, to be slices. But eventually we will want all to be slices. Might wait to add that until after I implement the error arena optimization.

Had to add const qualifier to certain tests to get everything to compile. Will make sure that is clean one the preceding PRs land.

Linux benchmark for ErrorCreate so far:
Originally: 338 ns
Part1: 327 ns
Part2: 127 ns
Part3: 95 ns (under 100!)

MacOs:
Originally: 1309 ns
Part1: 1204 ns
Part2: 319 ns
Part3: 160 ns