@ncteisen 2017-03-09
Closing this. I will redo this optimization on top of #9963, which has already switched to error's holding slices internally